Just require `python` during the build
As we need `python` on Windows and Unix to compile files and set the VC
correctly, simply just add it as a requirement without selectors. Also
specifying the `python` version for Windows in the `build` messes up
`conda-build-all` in determining the Windows build matrix, which in turn
messes up `conda-smithy`. So if we just leave this alone both
`conda-build-all` and `conda-smithy` should do their things and set this
correctly for the different build cases.