Test that libcxx is used
Simply add a test that will pass if libcxx is used, but fail otherwise.
Does this by checking that `_LIBCPP_VERSION` is defined during compile
time. If it is, then it creates a `main` function that passes otherwise
it creates one that fails. In either case, it prints out a message
informing either the value of `_LIBCPP_VERSION` or that
`_LIBCPP_VERSION` is undefined.