Dropped `-nostdlib` and system libraries
As part of the debugging process of this test, it appeared that we
needed the `-nostdlib` argument and, as a consequence, needed to
explicitly link `vcruntime`, `ucrt`, and `msvcrt`. However given various
changes since, it appears those arguments are no longer necessary. So we
are now dropping them. This should simplify the testing commands quite a
bit.